feat(client): callback for failed attribute read
Currently if reading an attribute asynchronously from the client
fails, the error is not reported.  Although it is logged, the calling
program is not informed as the callback is not called.  The program
may hang forever as the callbacks are used to make progress.  After
a call to __UA_Client_readAttribute_async() the callback should be
called in all cases.  So use NULL as value pointer in ValueAttributeRead()
to report failure.

To report the cause of the problem, the status code should be part
of the callback.  The generic UA_ClientAsyncServiceCallback callback
has a sufficient API as it reports a result.  There we have the
response header service result with the necessary information.  The
read attribute callbacks report a specfic type and need an additional
status code parameter.  Introduce a new UA_ClientAsyncAttributeCallback
type for these results.  To make the distincion clear, the existing
UA_ClientAsyncServiceCallback type passes an UA_Response instead
of a void pointer.  This is stricter helps to avoid some C casts.