fix(client) connect iterate without timeout
Commit ef4394b1144e845df93760b951ef0f6bef63d053 may have fixed the
connect timeout, but it broke async connect without timeout.  That
made it impossible to continue processing during TCP handshake.
While the connect operation is in progress, the select system call
will return immediately if the timeout is 0.  Do not close the
connection in this case, but allow the application to call run_iterate()
again.  For the tests in the CI this makes no difference as connect
over Linux loopback interface succeeds immediately.